Redefine login and api/access
